✏️ Split up Redux Actions
