+++ src/components/PlagueLogic.js
@@ -1,7 +1,7 @@
 import { Component } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
-import { infectPopulation } from '../redux/actions/actions'
+import { infectPopulation } from '../redux/actions/plague'
 
 class PlagueLogic extends Component {
   static propTypes = {
