+++ src/redux/actions/log.js
@@ -1,0 +1,14 @@
+export const ADD_LOG_ITEM = 'ADD_LOG_ITEM'
+
+/**
+ * Adds a log item to the Log.
+ * @param {Object} logItem LogItem details.
+ */
+export const addLogItem = logItem => {
+  return {
+    type: ADD_LOG_ITEM,
+    payload: {
+      logItem
+    }
+  }
+}
